+++ packages/docusaurus/src/pages/configuration.tsx
@@ -1,0 +1,7 @@
+import {Redirect} from '@docusaurus/router';
+import React      from 'react';
+
+// eslint-disable-next-line arca/no-default-export
+export default function Home() {
+  return <Redirect to={`/configuration/manifest`} />;
+}