+++ packages/plugin-nm/sources/NodeModulesLinker.ts
@@ -1,20 +1,20 @@
-import {structUtils, Report, Manifest, miscUtils, formatUtils}             from '@yarnpkg/core';
-import {Locator, Package, FinalizeInstallStatus, hashUtils}                from '@yarnpkg/core';
-import {Linker, LinkOptions, MinimalLinkOptions, LinkType}                 from '@yarnpkg/core';
-import {LocatorHash, Descriptor, DependencyMeta, Configuration}            from '@yarnpkg/core';
-import {MessageName, Project, FetchResult, Installer}                      from '@yarnpkg/core';
-import {PortablePath, npath, ppath, toFilename, Filename}                  from '@yarnpkg/fslib';
-import {VirtualFS, xfs, FakeFS, NativePath}                                from '@yarnpkg/fslib';
-import {ZipOpenFS}                                                         from '@yarnpkg/libzip';
-import {buildNodeModulesTree}                                              from '@yarnpkg/nm';
-import {NodeModulesLocatorMap, buildLocatorMap, NodeModulesHoistingLimits} from '@yarnpkg/nm';
-import {parseSyml}                                                         from '@yarnpkg/parsers';
-import {jsInstallUtils}                                                    from '@yarnpkg/plugin-pnp';
-import {PnpApi, PackageInformation}                                        from '@yarnpkg/pnp';
-import cmdShim                                                             from '@zkochan/cmd-shim';
-import {UsageError}                                                        from 'clipanion';
-import crypto                                                              from 'crypto';
-import fs                                                                  from 'fs';
+import {structUtils, Report, Manifest, miscUtils, formatUtils}              from '@yarnpkg/core';
+import {Locator, Package, FinalizeInstallStatus, hashUtils}                 from '@yarnpkg/core';
+import {Linker, LinkOptions, MinimalLinkOptions, LinkType, WindowsLinkType} from '@yarnpkg/core';
+import {LocatorHash, Descriptor, DependencyMeta, Configuration}             from '@yarnpkg/core';
+import {MessageName, Project, FetchResult, Installer}                       from '@yarnpkg/core';
+import {PortablePath, npath, ppath, toFilename, Filename}                   from '@yarnpkg/fslib';
+import {VirtualFS, xfs, FakeFS, NativePath}                                 from '@yarnpkg/fslib';
+import {ZipOpenFS}                                                          from '@yarnpkg/libzip';
+import {buildNodeModulesTree}                                               from '@yarnpkg/nm';
+import {NodeModulesLocatorMap, buildLocatorMap, NodeModulesHoistingLimits}  from '@yarnpkg/nm';
+import {parseSyml}                                                          from '@yarnpkg/parsers';
+import {jsInstallUtils}                                                     from '@yarnpkg/plugin-pnp';
+import {PnpApi, PackageInformation}                                         from '@yarnpkg/pnp';
+import cmdShim                                                              from '@zkochan/cmd-shim';
+import {UsageError}                                                         from 'clipanion';
+import crypto                                                               from 'crypto';
+import fs                                                                   from 'fs';
 
 const STATE_FILE_VERSION = 1;
 const NODE_MODULES = `node_modules` as Filename;
@@ -68,8 +68,11 @@ export class NodeModulesLinker implements Linker {
       throw err;
     }
 
-    const startingCwd = opts.project.configuration.startingCwd;
-    return locatorInfo.locations.find(location => ppath.contains(startingCwd, location)) || locatorInfo.locations[0];
+    // Sort locations from shallowest to deepest in terms of directory nesting
+    const sortedLocations = locatorInfo.locations.sort((loc1, loc2) => loc1.split(ppath.sep).length - loc2.split(ppath.sep).length);
+    // Find the location with shallowest directory nesting that starts inside node_modules of cwd
+    const startingCwdModules = ppath.join(opts.project.configuration.startingCwd, NODE_MODULES);
+    return sortedLocations.find(location => ppath.contains(startingCwdModules, location)) || locatorInfo.locations[0];
   }
 
   async findPackageLocator(location: PortablePath, opts: LinkOptions) {
@@ -472,7 +475,7 @@ async function findInstallState(project: Project, {unrollAliases = false}: {unro
   if (locatorState.__metadata.version > STATE_FILE_VERSION)
     return null;
 
-  const nmMode = locatorState.__metadata.nmMode || NodeModulesMode.HARDLINKS_LOCAL;
+  const nmMode = locatorState.__metadata.nmMode || NodeModulesMode.CLASSIC;
 
   const locatorMap: NodeModulesLocatorMap = new Map();
   const binSymlinks: BinSymlinkMap = new Map();
@@ -664,21 +667,24 @@ const buildLocationTree = (locatorMap: NodeModulesLocatorMap | null, {skipPrefix
   return locationTree;
 };
 
-const symlinkPromise = async (srcPath: PortablePath, dstPath: PortablePath) => {
-  let stats;
-
-  try {
-    if (process.platform === `win32`) {
+const symlinkPromise = async (srcPath: PortablePath, dstPath: PortablePath, windowsLinkType: WindowsLinkType) => {
+  // use junctions on windows if in classic mode
+  if (process.platform === `win32` && windowsLinkType === WindowsLinkType.JUNCTIONS) {
+    let stats;
+    try {
       stats = await xfs.lstatPromise(srcPath);
+    } catch (e) {
     }
-  } catch (e) {
+
+    if (!stats || stats.isDirectory()) {
+      await xfs.symlinkPromise(srcPath, dstPath, `junction`);
+      return;
+    }
+    // fall through to symlink
   }
 
-  if (process.platform == `win32` && (!stats || stats.isDirectory())) {
-    await xfs.symlinkPromise(srcPath, dstPath, `junction`);
-  } else {
-    await xfs.symlinkPromise(ppath.relative(ppath.dirname(dstPath), srcPath), dstPath);
-  }
+  // use symlink if tests for junction case fail
+  await xfs.symlinkPromise(ppath.relative(ppath.dirname(dstPath), srcPath), dstPath);
 };
 
 async function atomicFileWrite(tmpDir: PortablePath, dstPath: PortablePath, content: Buffer) {
@@ -774,7 +780,7 @@ type DirEntry = {
   symlinkTo: PortablePath;
 };
 
-const copyPromise = async (dstDir: PortablePath, srcDir: PortablePath, {baseFs, globalHardlinksStore, nmMode, packageChecksum}: {baseFs: FakeFS<PortablePath>, globalHardlinksStore: PortablePath | null, nmMode: {value: NodeModulesMode}, packageChecksum: string | null}) => {
+const copyPromise = async (dstDir: PortablePath, srcDir: PortablePath, {baseFs, globalHardlinksStore, nmMode, windowsLinkType: windowsLinkType, packageChecksum}: {baseFs: FakeFS<PortablePath>, globalHardlinksStore: PortablePath | null, nmMode: {value: NodeModulesMode}, windowsLinkType: WindowsLinkType, packageChecksum: string | null}) => {
   await xfs.mkdirPromise(dstDir, {recursive: true});
 
   const getEntriesRecursive = async (relativePath: PortablePath = PortablePath.dot): Promise<Map<PortablePath, DirEntry>> => {
@@ -837,7 +843,7 @@ const copyPromise = async (dstDir: PortablePath, srcDir: PortablePath, {baseFs, 
         mtimesChanged = true;
       }
     } else if (entry.kind === DirEntryKind.SYMLINK) {
-      await symlinkPromise(ppath.resolve(ppath.dirname(dstPath), entry.symlinkTo), dstPath);
+      await symlinkPromise(ppath.resolve(ppath.dirname(dstPath), entry.symlinkTo), dstPath, windowsLinkType);
     }
   }
 
@@ -1052,14 +1058,14 @@ async function persistNodeModules(preinstallState: InstallState, installState: N
   const locationTree = buildLocationTree(installState, {skipPrefix: project.cwd});
 
   const addQueue: Array<Promise<void>> = [];
-  const addModule = async ({srcDir, dstDir, linkType, globalHardlinksStore, nmMode, packageChecksum}: {srcDir: PortablePath, dstDir: PortablePath, linkType: LinkType, globalHardlinksStore: PortablePath | null, nmMode: {value: NodeModulesMode}, packageChecksum: string | null}) => {
+  const addModule = async ({srcDir, dstDir, linkType, globalHardlinksStore, nmMode, windowsLinkType, packageChecksum}: {srcDir: PortablePath, dstDir: PortablePath, linkType: LinkType, globalHardlinksStore: PortablePath | null, nmMode: {value: NodeModulesMode},  windowsLinkType: WindowsLinkType, packageChecksum: string | null}) => {
     const promise: Promise<any> = (async () => {
       try {
         if (linkType === LinkType.SOFT) {
           await xfs.mkdirPromise(ppath.dirname(dstDir), {recursive: true});
-          await symlinkPromise(ppath.resolve(srcDir), dstDir);
+          await symlinkPromise(ppath.resolve(srcDir), dstDir, windowsLinkType);
         } else {
-          await copyPromise(dstDir, srcDir, {baseFs, globalHardlinksStore, nmMode, packageChecksum});
+          await copyPromise(dstDir, srcDir, {baseFs, globalHardlinksStore, nmMode, windowsLinkType, packageChecksum});
         }
       } catch (e) {
         e.message = `While persisting ${srcDir} -> ${dstDir} ${e.message}`;
@@ -1270,6 +1276,7 @@ async function persistNodeModules(preinstallState: InstallState, installState: N
   const reportedProgress = report.reportProgress(progress);
   const nmModeSetting = project.configuration.get(`nmMode`);
   const nmMode = {value: nmModeSetting};
+  const windowsLinkType = project.configuration.get(`winLinkType`) as WindowsLinkType;
 
   try {
     // For the first pass we'll only want to install a single copy for each
@@ -1288,7 +1295,7 @@ async function persistNodeModules(preinstallState: InstallState, installState: N
     for (const entry of addList) {
       if (entry.linkType === LinkType.SOFT || !persistedLocations.has(entry.srcDir)) {
         persistedLocations.set(entry.srcDir, entry.dstDir);
-        await addModule({...entry, globalHardlinksStore, nmMode, packageChecksum: realLocatorChecksums.get(entry.realLocatorHash) || null});
+        await addModule({...entry, globalHardlinksStore, nmMode, windowsLinkType, packageChecksum: realLocatorChecksums.get(entry.realLocatorHash) || null});
       }
     }
 
@@ -1308,7 +1315,7 @@ async function persistNodeModules(preinstallState: InstallState, installState: N
     await xfs.mkdirPromise(rootNmDirPath, {recursive: true});
 
     const binSymlinks = await createBinSymlinkMap(installState, locationTree, project.cwd, {loadManifest});
-    await persistBinSymlinks(prevBinSymlinks, binSymlinks, project.cwd);
+    await persistBinSymlinks(prevBinSymlinks, binSymlinks, project.cwd, windowsLinkType);
 
     await writeInstallState(project, installState, binSymlinks, nmMode, {installChangedByUser});
 
@@ -1320,7 +1327,7 @@ async function persistNodeModules(preinstallState: InstallState, installState: N
   }
 }
 
-async function persistBinSymlinks(previousBinSymlinks: BinSymlinkMap, binSymlinks: BinSymlinkMap, projectCwd: PortablePath) {
+async function persistBinSymlinks(previousBinSymlinks: BinSymlinkMap, binSymlinks: BinSymlinkMap, projectCwd: PortablePath, windowsLinkType: WindowsLinkType) {
   // Delete outdated .bin folders
   for (const location of previousBinSymlinks.keys()) {
     if (ppath.contains(projectCwd, location) === null)
@@ -1358,7 +1365,7 @@ async function persistBinSymlinks(previousBinSymlinks: BinSymlinkMap, binSymlink
         await cmdShim(npath.fromPortablePath(target), npath.fromPortablePath(symlinkPath), {createPwshFile: false});
       } else {
         await xfs.removePromise(symlinkPath);
-        await symlinkPromise(target, symlinkPath);
+        await symlinkPromise(target, symlinkPath, windowsLinkType);
         if (ppath.contains(projectCwd, await xfs.realpathPromise(target)) !== null) {
           await xfs.chmodPromise(target, 0o755);
         }