+++ packages/plugin-npm/sources/npmHttpUtils.ts
@@ -1,11 +1,11 @@
-import {Configuration, Ident, formatUtils, httpUtils} from '@yarnpkg/core';
-import {MessageName, ReportError}                     from '@yarnpkg/core';
-import {prompt}                                       from 'enquirer';
-import {URL}                                          from 'url';
+import {Configuration, Ident, formatUtils, httpUtils, nodeUtils, StreamReport} from '@yarnpkg/core';
+import {MessageName, ReportError}                                              from '@yarnpkg/core';
+import {prompt}                                                                from 'enquirer';
+import {URL}                                                                   from 'url';
 
-import {Hooks}                                        from './index';
-import * as npmConfigUtils                            from './npmConfigUtils';
-import {MapLike}                                      from './npmConfigUtils';
+import {Hooks}                                                                 from './index';
+import * as npmConfigUtils                                                     from './npmConfigUtils';
+import {MapLike}                                                               from './npmConfigUtils';
 
 export enum AuthType {
   NO_AUTH,
@@ -108,7 +108,7 @@ export async function post(path: string, body: httpUtils.Body, {attemptedAs, con
       throw error;
     }
 
-    otp = await askForOtp();
+    otp = await askForOtp(error, {configuration});
     const headersWithOtp = {...headers, ...getOtpHeaders(otp)};
 
     // Retrying request with OTP
@@ -144,7 +144,7 @@ export async function put(path: string, body: httpUtils.Body, {attemptedAs, conf
       throw error;
     }
 
-    otp = await askForOtp();
+    otp = await askForOtp(error, {configuration});
     const headersWithOtp = {...headers, ...getOtpHeaders(otp)};
 
     // Retrying request with OTP
@@ -180,7 +180,7 @@ export async function del(path: string, {attemptedAs, configuration, headers, id
       throw error;
     }
 
-    otp = await askForOtp();
+    otp = await askForOtp(error, {configuration});
     const headersWithOtp = {...headers, ...getOtpHeaders(otp)};
 
     // Retrying request with OTP
@@ -259,9 +259,44 @@ async function whoami(registry: string, headers: {[key: string]: string} | undef
   }
 }
 
-async function askForOtp() {
-  if (process.env.TEST_ENV)
-    return process.env.TEST_NPM_2FA_TOKEN || ``;
+async function askForOtp(error: any, {configuration}: {configuration: Configuration}) {
+  const notice = error.originalError?.response.headers[`npm-notice`];
+
+  if (notice) {
+    await StreamReport.start({
+      configuration,
+      stdout: process.stdout,
+      includeFooter: false,
+    }, async report => {
+      report.reportInfo(MessageName.UNNAMED, notice.replace(/(https?:\/\/\S+)/g, formatUtils.pretty(configuration, `$1`, formatUtils.Type.URL)));
+
+      if (!process.env.YARN_IS_TEST_ENV) {
+        const autoOpen = notice.match(/open (https?:\/\/\S+)/i);
+        if (autoOpen && nodeUtils.openUrl) {
+          const {openNow} = await prompt<{openNow: boolean}>({
+            type: `confirm`,
+            name: `openNow`,
+            message: `Do you want to try to open this url now?`,
+            required: true,
+            initial: true,
+            onCancel: () => process.exit(130),
+          });
+
+          if (openNow) {
+            if (!await nodeUtils.openUrl(autoOpen[1])) {
+              report.reportSeparator();
+              report.reportWarning(MessageName.UNNAMED, `We failed to automatically open the url; you'll have to open it yourself in your browser of choice.`);
+            }
+          }
+        }
+      }
+    });
+
+    process.stdout.write(`\n`);
+  }
+
+  if (process.env.YARN_IS_TEST_ENV)
+    return process.env.YARN_INJECT_NPM_2FA_TOKEN || ``;
 
   const {otp} = await prompt<{otp: string}>({
     type: `password`,
@@ -270,6 +305,8 @@ async function askForOtp() {
     required: true,
     onCancel: () => process.exit(130),
   });
+
+  process.stdout.write(`\n`);
 
   return otp;
 }