+++ packages/yarnpkg-core/sources/MessageName.ts
@@ -94,6 +94,7 @@ export enum MessageName {
   NETWORK_DISABLED = 80,
   NETWORK_UNSAFE_HTTP = 81,
   RESOLUTION_FAILED = 82,
+  AUTOMERGE_GIT_ERROR = 83,
 }
 
 export function stringifyMessageName(name: MessageName | number): string {