+++ packages/yarnpkg-fslib/sources/patchFs/FileHandle.ts
@@ -1,4 +1,5 @@
 import type {BigIntStats, ReadStream, StatOptions, Stats, WriteStream, WriteVResult} from 'fs';
+import {createInterface}                                                             from 'readline';
 
 import type {CreateReadStreamOptions, CreateWriteStreamOptions, FakeFS}              from '../FakeFS';
 import type {Path}                                                                   from '../path';
@@ -210,6 +211,13 @@ export class FileHandle<P extends Path> {
     } finally {
       this[kUnref]();
     }
+  }
+
+  readLines(options?: CreateReadStreamOptions) {
+    return createInterface({
+      input: this.createReadStream(options),
+      crlfDelay: Infinity,
+    });
   }
 
   stat(