+++ packages/yarnpkg-pnp/sources/esm-loader/loaderFlags.ts
@@ -8,3 +8,6 @@ export const HAS_UNFLAGGED_JSON_MODULES = major > 17 || (major === 17 && minor >
 
 // JSON modules requires import assertions after https://github.com/nodejs/node/pull/40250
 export const HAS_JSON_IMPORT_ASSERTION_REQUIREMENT = major > 17 || (major === 17 && minor >= 1) || (major === 16 && minor > 14);
+
+// The message switched to using an array in https://github.com/nodejs/node/pull/45348
+export const WATCH_MODE_MESSAGE_USES_ARRAYS = major > 19 || (major === 19 && minor >= 2);